backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pillow==10.3.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
